refactor: 메뉴 선택 컨트롤러 리팩토링
