+++ src/main/java/com/dev/voucherproject/controller/console/MenuController.java
@@ -1,5 +1,6 @@
 package com.dev.voucherproject.controller.console;
 
+
 @FunctionalInterface
 public interface MenuController {
     void execute();
