+++ src/main/java/com/dev/voucherproject/controller/console/MenuUsingConsoleAndStorage.java
@@ -1,13 +1,13 @@
 package com.dev.voucherproject.controller.console;
 
-import com.dev.voucherproject.storage.VoucherStorage;
+import com.dev.voucherproject.storage.voucher.VoucherStorage;
 import com.dev.voucherproject.view.Console;
 
 public abstract class MenuUsingConsoleAndStorage implements MenuController {
     protected final VoucherStorage voucherStorage;
     protected final Console console;
 
-    public MenuUsingConsoleAndStorage(VoucherStorage voucherStorage, Console console) {
+    protected MenuUsingConsoleAndStorage(VoucherStorage voucherStorage, Console console) {
         this.voucherStorage = voucherStorage;
         this.console = console;
     }
